Crate lazy_regex_proc_macros

Macros

lazy_regex

Return an instance of once_cell::sync::Lazy<regex::Regex> or once_cell::sync::Lazy<regex::bytes::Regex> that you can use in a public static declaration.

regex

Return a lazy static Regex checked at compilation time and built at first use.

regex_captures

Extract captured groups as a tuple of &str.

regex_find

Extract the leftmost match of the regex in the second argument, as a &str, or a &[u8] if the B flag is set.

regex_is_match

Test whether an expression matches a lazy static regular expression (the regex is checked at compile time)

regex_replace

Replaces the leftmost match in the second argument with the value returned by the closure given as third argument.

regex_replace_all

Replaces all non-overlapping matches in the second argument with the value returned by the closure given as third argument.